create new component structure
