+++ client/app/auth/logout/LogoutContainer.js
