--- client/components/pages/DecksPage.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import PageTemplate from '../templates/PageTemplate';
-import DecksContainer from '../organisms/DecksContainer';
-
-const ItemsPage = () => {
-  return (
-    <PageTemplate>
-      <DecksContainer />
-    </PageTemplate>
-  );
-};
-
-export default ItemsPage;
