--- client/components/pages/ItemsPage.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-import PageTemplate from '../templates/PageTemplate';
-import ItemsContainer from '../organisms/ItemsContainer';
-
-const ItemsPage = () => {
-  return (
-    <PageTemplate>
-      <ItemsContainer />
-    </PageTemplate>
-  );
-};
-
-export default ItemsPage;
